feat: add ACL

+++ src/controllers/v1.0/ListController.ts
@@ -1,7 +1,8 @@
-import {NextFunction, Request, Response} from 'express';
+import {NextFunction, Response} from 'express';
 
-import {HttpStatus} from '../../utils/appError';
+import {AppError, ErrorCode, HttpStatus} from '../../utils/appError';
 import {formatList, formatSubTask} from '../../utils/list';
+import ac from '../../utils/ac';
 
 import Controller, {Methods} from '../../core/Controller';
 
@@ -58,7 +59,7 @@ class ListController extends Controller {
     },
     {
       path: '/:listID/subtasks/:subtaskID',
-      method: Methods.PUT,
+      method: Methods.DELETE,
       handler: this.destroySubTask,
       localMiddleware: [authMiddleware],
     },
@@ -70,6 +71,12 @@ class ListController extends Controller {
 
   async create(req: any, res: Response, next: NextFunction) {
     try {
+      const permission = ac.can(req.user.role).createOwn('list');
+
+      if (!permission.granted) {
+        throw new AppError(ErrorCode.List.ForBidden);
+      }
+
       const {name, description, startAt, endAt, color} = req.body;
 
       const list = await ListService.create({
@@ -91,8 +98,14 @@ class ListController extends Controller {
     }
   }
 
-  async show(req: Request, res: Response, next: NextFunction) {
+  async show(req: any, res: Response, next: NextFunction) {
     try {
+      const permission = ac.can(req.user.role).readOwn('list');
+
+      if (!permission.granted) {
+        throw new AppError(ErrorCode.List.ForBidden);
+      }
+
       const list = await ListService.show({
         id: req.params.id,
       });
@@ -102,13 +115,21 @@ class ListController extends Controller {
           data: formatList(list),
         });
       }
+
+      throw new AppError(ErrorCode.List.ForBidden);
     } catch (error) {
       next(error);
     }
   }
 
-  async update(req: Request, res: Response, next: NextFunction) {
+  async update(req: any, res: Response, next: NextFunction) {
     try {
+      const permission = ac.can(req.user.role).updateOwn('list');
+
+      if (!permission.granted) {
+        throw new AppError(ErrorCode.List.ForBidden);
+      }
+
       const {name, description, startAt, endAt, color} = req.body;
 
       await ListService.update({
@@ -120,14 +141,20 @@ class ListController extends Controller {
         color,
       });
 
-      res.status(HttpStatus.Ok).send();
+      res.status(HttpStatus.NoContent).send();
     } catch (error) {
       next(error);
     }
   }
 
-  async index(req: Request, res: Response, next: NextFunction) {
+  async index(req: any, res: Response, next: NextFunction) {
     try {
+      const permission = ac.can(req.user.role).readOwn('list');
+
+      if (!permission.granted) {
+        throw new AppError(ErrorCode.List.ForBidden);
+      }
+
       const {limit, page, name, description, startAt, endAt} = req.query;
       const _limit = limit ? parseInt(limit as string) : 10;
       const _page = page ? parseInt(page as string) : 1;
@@ -154,6 +181,12 @@ class ListController extends Controller {
 
   async createSubTask(req: any, res: Response, next: NextFunction) {
     try {
+      const permission = ac.can(req.user.role).createOwn('subtask');
+
+      if (!permission.granted) {
+        throw new AppError(ErrorCode.SubTask.ForBidden);
+      }
+
       const {name, description} = req.body;
 
       const subTask = await SubTaskService.create({
@@ -172,8 +205,14 @@ class ListController extends Controller {
     }
   }
 
-  async showSubTask(req: Request, res: Response, next: NextFunction) {
+  async showSubTask(req: any, res: Response, next: NextFunction) {
     try {
+      const permission = ac.can(req.user.role).readOwn('subtask');
+
+      if (!permission.granted) {
+        throw new AppError(ErrorCode.SubTask.ForBidden);
+      }
+
       const subTask = await SubTaskService.show({
         id: req.params.subtaskID,
         listID: req.params.listID,
@@ -189,8 +228,14 @@ class ListController extends Controller {
     }
   }
 
-  async updateSubTask(req: Request, res: Response, next: NextFunction) {
+  async updateSubTask(req: any, res: Response, next: NextFunction) {
     try {
+      const permission = ac.can(req.user.role).updateOwn('subtask');
+
+      if (!permission.granted) {
+        throw new AppError(ErrorCode.SubTask.ForBidden);
+      }
+
       const {name, description} = req.body;
 
       await SubTaskService.update({
@@ -206,8 +251,14 @@ class ListController extends Controller {
     }
   }
 
-  async destroySubTask(req: Request, res: Response, next: NextFunction) {
+  async destroySubTask(req: any, res: Response, next: NextFunction) {
     try {
+      const permission = ac.can(req.user.role).deleteOwn('subtask');
+
+      if (!permission.granted) {
+        throw new AppError(ErrorCode.SubTask.ForBidden);
+      }
+
       await SubTaskService.destroy({
         id: req.params.subtaskID,
         listID: req.params.listID,
